requirements.txt
flask>=3.0.0
flask-login>=0.6.3
flask-sqlalchemy>=3.1.1
sqlalchemy>=2.0.23
pandas>=2.1.4
openpyxl>=3.1.2
xlrd>=2.0.1
qrcode[pil]>=7.4.2
Pillow>=10.1.0
pyzbar>=0.1.9
fpdf2>=2.7.6
werkzeug>=3.0.0

